oaaclient class changes
Moved `OAAPermission` and `OAAIdentityType` enums from `client.py` to `templates.py`
to avoid future circular import issues.

Updated import statements for existing connectors to reflect change.
